# Docs/openapi.yaml
openapi: 3.0.3
info:
  title: tracklist API
  description: |
    REST API для автентифікації, керування користувачами,
    медіа-контентом, відгуками, коментарями, списками та взаємодіями.
  version: 1.0.0
servers:
  - url: http://localhost:7288
    description: Local ASP.NET Core API
...
paths:
...
  /api/user/{email}:
    get:
      tags: [Users]
      summary: Отримання профілю користувача
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              example:
                data:
                  email: user@mail.com
                  role: User
        "400":
          description: User not found

    delete:
      tags: [Users]
      summary: Деактивація (видалення) профілю користувача
      responses:
        "200":
          description: User deleted
        "400":
          description: User not found
...